[API CHANGE] related tdf#85592: drop duplicate use of .uno:CharBackgroundExt
The primary purpose of CharBackgroundExt is a bool flag to
indicate that the watering-can for text background coloring
is active.
It could also function as a synonym for BackColor in the "execute"
state. A prior (NFC) commit removed internal LO use in this way,
so lets drop the ability to do this altogether.

I'm not sure why anyone would have selected to use CharBackgroundExt
instead of .uno:BackColor in API calls. Perhaps they discovered it via
LO_COLLECT_UIINFO. A google search didn't indicate any significant
use or suggestion to use CharBackgroundExt.

P.S. I'm also trying to deprecate BackColor and replace it with
.uno:CharBackColor - which is what this bug report is actually about.

Change-Id: I8d8a2aedb1af18e4a10ff09ad0dabc58e90514a8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138167
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>