tdf#103492: Messed Arabic letter spacing in text starting with LTR character(s)
This partially reverts the still relevant parts of commit

commit 4682b934a28d4ffab0c9615f365d274ad942585a
Author: Martin Hosken <martin_hosken@sil.org>
Date:   Thu Sep 10 10:14:18 2015 +0700

    Refactor graphite integration and update graphite

    Change-Id: I09db7262c2a6180a078620ba086dd7a4c132bd82
    Reviewed-on: https://gerrit.libreoffice.org/18459
    Tested-by: Jenkins <ci@libreoffice.org>
    Reviewed-by: László Németh <nemeth@numbertext.org>
    Reviewed-by: Martin Hosken <martin_hosken@sil.org>

No mention of the commit about this change or its relevance to Graphite
integration. But the separate Graphite layout engine is long gone and
Graphite handling goes through HarfBuzz now, so whatever the change was
for is probably no longer relevant.

I reverted only one the two changes in impedit3.cxx, as the other one
does not seem to make any difference for this issue.

Change-Id: Id70a711d6231cce52c9a32dbc3604860b1da9f4a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138217
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>