tdf#144092 pptx import: fix import of empty cell and shape text properties
For table cells and shapes without any text, text properties weren't
correctly imported. Both the master styles and properties defined in
the endParaRPr were ignored.

It appears the current implementation does attempt to import those in
TextParagraph::insertAt(...), after the attempt they end up in the
TextRange('s nodes) but they are ignored.

This commit implements a fix by: If the cell or shape has no text, the
text properties are also pushed into the shape or cell's propset
directly

Change-Id: Ie425b3e9a73937a2503e47322e25f37ed9a2a6be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138651
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>