tdf#53530 Only show 1 error dialog instead of 3 on export to PDF
SfxObjectShell::ExecFile_Impl is called when exporting to PDF, and
the 3rd error message pops up here, so we just add a check to see
if we're about to show the general IO error during pdf export and
skip showing the error message there.

The first 2 error messages pop up from
SfxMedium::TransactedTransferForFS_Impl, and so we just add a check
to see if the first error message popped up in which case an abort
error code is returned and we then skip popping up the 2nd error.

Change-Id: Ifa85ca1e451b1f87c6ddd89a98d377ea04aeaa0f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138333
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>