ScriptForge - (Calc) redesign CompactUp() and CompactLeft()
This commit is mainly about the introduction of
a new internal method:
   _ComputeFilter(range, filterformula, filterscope)

A FilterFormula is a Calc formula that returns TRUE or FALSE
   The formula is expressed in terms of
   - the top-left cell of the range when FilterScope = "CELL"
   - the topmost row of the range when FilterScope = "ROW"
   - the leftmost column of the range when FilterScope = "COLUMN"
   After pasting, the relative and absolute references
   will be interpreted correctly.
The FilterScope indicates the way the formula is applied,
once by row, column or individual cell.

The concept of FilterFormula was already used by CompactUp()
and CompactLeft(). Their implicit (Filter)scopes were resp.
"ROW" and "COLUMN".

The _ComputeFilter() method returns an array of subranges
contained in the initial range that match the filter.

The isolation the code for the management of filters
applied on ranges makes the later use of the concepts of
FilterFormula and FilterScope reusable for other methods.

CompactUp() and CompactLeft() are functionally unchanged.
No impact on documentation.

Change-Id: I7c4e890b54f315486f29b5434a3c236167e2f9ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138368
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins