tdf#150457 Fix crash on hovering removed footnote reference
The regression was caused by aa6bbdd134c509d2c96148222d6cf862ec12a2c6.
Before that, c491323e2d1c64014ed543a94f335e8cee673171 changed the cursor
jumping behavior to fix tdf#81226. The goal was to make LibreOffice
behave as described in the description of the issue:

"When your cursor is on the last line but not at the end of it, simply
press the down arrow to go at the end of the file (end of that line)."

The same behavior was achieved for pressing down on the last line.

The above patches allowed removing the footnotes and their numbers
(except the first number) while the reference were still present. Then,
on hovering a removed footnote reference, a crash was happening. It
should be noted that even with this patch, removing the first reference
number was not possible.

This fix limits this behavior to anywhere other than inside a footnote.
In this way, it prevents crashing.

With this fix, the behavior of the cursor goes back to what it was
before c491323e2d1c64014ed543a94f335e8cee673171. The user will not be
able to select all the footnotes at once. Also, removing footnote number
will not be possible.

It is worth noting that not being able to select all the footnotes at
once prevents the user from changing the characteristics of the
footnotes all at the same time. Also, not being able to remove the
footnote numbers, prevents the user from having footnotes with custom
numbers written manually. Both of these actions are possible in MS Word.
The other difference is that by going up, one can go out of the footnote
section in MS Word, but this is not the case for LibreOffice.

A UI test is added to make sure the crash will not happen again. The
test can be run by:

 cd sw && make -srj1 UITest_writer_tests2 \
  UITEST_TEST_NAME="deleteFootnotes.tdf150457.test_delete_footnotes" \
  SAL_USE_VCLPLUGIN=gen

Change-Id: I1ce7d2189355f6763b6f31219e00516ff7a8164e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138562
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>