Don't position dialogs relative to other dialogs
When calling setPosSize on a dialog, and there are other dialogs on top of it,
the dialog was positioned relative to its parent dialog.

Fix this, so that dialogs are always positioned relative to the main window (if there is one).

Revert of/Follow up to 2d10767f0a9d7595e042f7ce23a81bcd947199fe

Change-Id: I69f189865b118f9bc20077d48591fbd9e83b014f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138994
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@allotropia.de>