crashtesting: fix DOCX export of forum-mso-de-102317.docx
This went wrong in d6654a07da0e14bf86150ea3ec4007900ae61a45 (DOCX
import: handle SDT around citation field, 2014-07-16), the problem is
that a run/inline SDT is pending, but we wrote a header stream in the
meantime, so the invariant that Start/EndParagraph() in
DocxAttributeOutput is called in pairs doesn't help, which results in
not-well-formed XML, which is an assertion failure in debug build.

Fix the problem similar to b846826ca53bcf101d76d716d6493db87bb0803d (sw:
refactoring for docx sdt blocks output, 2021-11-16), which already did
the same for paragraph/block SDTs.

Change-Id: I5e7273b839bb5c1f841c0fab9b68b4ca239cca18
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139089
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins