tdf#139627 Test justified Arabic/Persian text to avoid gaps/big overlaps
With the patches cf8ff7e..e64b643 from Khaled, many problems with the
justified Arabic/Persian text which were related to Kashida are fixed.
Here, we add a test for tdf#139627 which converts the example file to
the PDF format, then checks the width and the position of the characters
in the output to make sure:

* The characters are present in the PDF file in the correct order
* The characters are joined together
* The diacritic mark is positioned correctly
* The overlapping of the tatweel character is not more than 10% of the
  first character

Sample odt file uses "Noto Arabic Sans" font, which is available via
LibreOffice on all platforms.

One may run the test with:

    make CPPUNIT_TEST_NAME=testTdf139627 -sr CppunitTest_vcl_pdfexport

Change-Id: I7a826a1b43ee842978decb0cf9a5e2a3b7219982
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138328
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>