tdf#150756 - Assertion when opening report for editing
regression from
   commit 3f1a74ed1036d9dd22d631abead8f8fbbaf98965
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Thu May 27 10:27:46 2021 +0200
    ref-count SdrObject
this logic is not meant to kick in for reportdesigner.

If this commit causes further regressions, and we need to reinstate some
form of this logic, it should be conditional on a flag which is
deliberately set by the modules (probably only sd) which actually need
it.
And the flag should be an explicit "I want to kill attached UNO shapes
when the page changes" rather than being conditional on some other
condition which is, at best, only tangentially related.

Change-Id: Ibc09149ab4ff88fde8d1bd1e0545e255ccdba50a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139520
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>