tdf#149547 fix "tdf#75297 sw uno: override default num char style"
The problem was that things like chapter numbering were
creating this style on import.
Interesting that it didn't do it for my use case.

This fixes 7.5 commit 7b96c6e33034c14fd45fd4ed68f136e6ef8af0f9.

Originally I left this clause "as is" because I had not yet removed
the obsolete part of the condition. But now I can change anything
in that .isEmpty() clause since
commit 5564517810f2abd4bd20a09c074e0ec7ec66577b
    remove unused sw UNO_NAME_CHARACTER_FORMAT_NONE

Change-Id: Ie52a0d1392d49ba6ae8a5620ab453ff7129bce7b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139584
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>