tdf#148214 RTF import: avoid fake paragraph for \page when possible
The bugdoc has 2 pages, an explicit page break between them and the
first page has enough content that an additional fake paragraph at the
end would lead to a 3rd, unwanted page.

The fake paragraph was introduced in commit
0b321d09c1d7b21bac745f52429fe822d7f6c21e (fdo#48104 fix RTF import of
\page in inner groups, 2012-04-02), because dmapper ignores more than 1
page breaks in a paragraph.

Fix the problem by only inserting the fake paragraph in case the "page
break before" paragraph property would have no effect (very first
paragraph in the document) or we already sent characters (para props are
lazy-sent on the first character). This keeps existing cases working,
but avoids the unwanted fake paragraph in the bugdoc.

I suspect the root cause is that dmapper doesn't handle multiple page
breaks in a paragraph -- once that's fixed, this parBreak() call can be
completely removed.

Adjusted tests:

- testImportHeaderFooter: this asserted the presence of fake paragraphs,
  which are now gone in the easy case, so this change is an improvement

- testTdf133437: no visual difference before/after, probably the test
  intent was just "make sure this page has several shapes", the exact
  number isn't that interesting

Change-Id: Idd2b8a70b4122eb08d9d305018d384dc0bbb276a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139704
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>