tdf#140302 RTF import: fix unexpected repeated image
The bugdoc has 2 images, but the resulting doc model had 3 images since
commit efb2a20a9c3dead0b06069919ad34a65371bd9dc (tdf#81943 sw RTF
import: fix missing wrap in background for in-table shape, 2018-07-16).

The trouble was that shape replay wants to access the current shape at
m_aStates.top().aShape, but we only set it, forgot to unset it after
replay was done, which can result in duplicated insertions.

Fix this by saving the old current shape on the stack, set the current
shape to the wanted value and then undo it after replay.

This only affects documents that have a \picprop group, since the
unwanted insertion happened when popping that group.

Change-Id: I088054de2332deb4b45f7e269672fa799dd49c2f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139818
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>