SwNavigator: Make Drawing objects context menu entry 'Edit..' edit
This patch uses SwEditWin::KeyInput handling of KEY_RETURN for the
Drawing objects context menu entry 'Edit...' which moves focus of the
selected drawing object into the text box of the drawing object.
Previous to this patch the 'Edit...' menu entry would select the
drawing object the same as the 'Go to' menu entry.

Change-Id: I4d6d278e25badd086dc718c4c7ea02ce6330a45d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140037
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>