tdf#149551 unit test writing mode in shape in Impress
Commit 31a02c351dd997a394b305d1fb747e8da68cef2e has among other things
enabled, that a shape in Impress can use a writing mode corresponding to
vert='vert' in a pptx file. It is new value text::WritingMode2::TB_RL90
for property WritingMode. It is written with value 'tb-rl90' to attribute
loext:writing-mode in ODF.

The same commit has enabled, that the writing mode 'bt-rl' can be used
not only for frames in Writer but for shapes in Impress too.

The unit tests here test round-trip of these property values.

Change-Id: I8dcf9e234d25c2769ed60edd0e435c3606f368bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140066
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>