tdf#151051 Fix promote/demote outline-level promotes/demotes both
outline nodes for outline nodes with table content as last content
followed immediately by an outline node

SwCursorShell::MakeOutlineSel function, currently used to select the
range of outline nodes for outline-level promote/demote, makes a
cursor selection of outline nodes to the start of the next outline
node past the outline nodes to be promoted/demoted. The selection is
then adjusted to the end content of the previous node. When the
previous node is a table, the selection does not move into the table,
but remains at the start of outline node that is one node after the
wanted range. This patch avoids use of the MakeOutlineSel function
for outline-level promote/demote.

Change-Id: Ida505df6ec172681868e86570ae523494ed7929e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141128
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>