Simplify a bit by using sal_uInt16 in vcl/BitmapPopArtFilter
Since we're in the "if (bRet)" block, it means
 22     bool bRet = isPalettePixelFormat(aBitmap.getPixelFormat())
 23                 || aBitmap.Convert(BmpConversion::N8BitColors);
is verified.

isPalettePixelFormat implementation is:
29  constexpr bool isPalettePixelFormat(PixelFormat ePixelFormat)
30  {
31      assert(ePixelFormat != PixelFormat::INVALID);
32      return sal_uInt16(ePixelFormat) <= 8;
33  }

So we know we're using 8 bits max and this line:
pWriteAcc->GetBitCount()
can't give more than 8 and we can safely declare nEntryCount as sal_uInt16 (idem for "n" just below)

Since "nFirstEntry" and "nLastEntry" are related to "nEntryCount", idem for mnIndex
they can also be sal_uInt16.

Thanks to these, we can avoid all sal::static_int_cast<sal_uInt16> conversions.

Change-Id: I8cac2d01f00be33c86058c7a6eb7b9e25fb2635e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140206
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>