tdf#150990 ODT export: fix zero layout size of scale/scale images
The bugdoc has a broken image where both axis are set to a relative size
of 255%, which means "keep aspect", and the expectation is that only one
axis uses this to match the size of the other axis.

The problem was that commit d210218ba2b1adf1bebb20d1ad0df34edd71edba
(ODT export: fix fallback svg:width/height for text frames with relative
sizes, 2022-03-16) assumed that only one axis uses "scale", so we always
get a non-zero layout size.

Fix the problem by only using the layout size when exactly one axis uses
the "scale" percentage, which is the expected the case.

This way we just keep the aspect ratio of broken documents, then the UI
will in practice ignore the scale request of the width.

Change-Id: I2dbd6bb92f912f6185422fd301b99b284a66ef74
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140218
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins